fix(solver): relate generic Pick/Omit mapped source to index-signature target
A generic homomorphic mapped type `{ [P in K]: Src[P] }` (the instantiated
body of `Pick<Src, K>` / `Omit<Src, K>` and user aliases of that shape) was
reported as not assignable to an index-signature target `{ [x: string]: V }`
when `Src` is a type parameter constrained to an index-signature object and the
key set `K` is a generic subset of `keyof Src`. The per-key access `Src[P]`
stays deferred because `P`'s key constraint resolves to a generic `keyof T`
(when the alias parameter is unconstrained, as `Pick`/`Omit` declare it), so
`check_generic_mapped_to_index_target` could not relate the template to the
target index value and produced a false TS2322 plus a follow-on TS2345.

tsc treats every property value of such a mapped type as a value drawn from the
homomorphic source. Add a fall-back: when the direct template subtype check
fails and the template is `Src[P]` (indexed by the mapped's own iteration
parameter), relate every value `Src`'s apparent type can yield — its declared
property values and string/number index-signature values — to the target index
value. The mapped type is assignable iff all of them are assignable.

Also split the one-sided Mapped relation cluster out of `generics.rs` into a new
`mapped_target` module to stay under the file-size cap.

Goal: green (superstruct canary row)
